دریافت لیست سفارشات در صفحه اصلی

+++ applibrary/src/main/java/com/hoomanholding/applibrary/model/data/request/OrderToggleStateRequest.kt
@@ -10,6 +10,6 @@ import com.hoomanholding.applibrary.model.data.enums.EnumState
 data class OrderToggleStateRequest(
     val state: EnumState,
     val description: String?,
-    val orderIds: List<Int>,
+    val orderIds: List<Long>,
     val rejectReasonId: Int?
 )
